services/flights_with_weather/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Ramane23 <contactsouley@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10 <=3.12.5"
cdsapi = "^0.7.1"
loguru = "^0.7.2"
pydantic = ">=2.7,<2.8"
quixstreams = "^2.9.0"
ecmwf-api-client = "^1.6.3"
xarray = "^2024.7.0"
setuptools = "^73.0.1"
pygrib = "^2.1.6"
geopy = "^2.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
